Revert adding epicblanchenoir
This theme simply just swap the graphics from epic-noir to be black and white without adjusting it to rg351p aspect ratio. Many systems are missing and some of them have a bad names.  This theme should be consider as a WIP, instead of adding it to package as a ready to use one