Introduce beta releases (pre-release in github) which can be used to update from the 'beta' channel on the device.
Mostly the changes are around building the beta builds themselves.

There is a small change to the update script to ensure that a 'release'
build will always take precendence over a 'beta' build so if you switch
back to the 'release' channel you can update back to the released
build.