Initialize audio on boot
Since we've switched to 'softvol' alsa plugin to adjust volume in
software (since 552 does not appear to have a sound card which can be
adjusted by this kernel), we need to initialize the sound card so the
'Playback' mixer exists and volume can be changed.

Without this initialization, volume cannot be changed until sound is
played.

To do this, we play a .001 second, totally silent .wav file.