Fix mali version mismatch
mali-midgard is for RG552 and mali-bifrost is for RG351.  If they are set mixed/crossed, it will
lead to issues loading any UI (emulation station, etc) as they have conflicting versions of
/usr/lib/libmali.so.1.9.0

There is a variable - $OPENGLES which contains the appropriate version for the device, that should be used in place of 'mali-midgard' or 'mali-bifrost' to avoid if/else statments.

This change just updates any mali-midgard/mali-bifrost dependencies to use $OPENGLES so the incorrect mali version is not pulled in.