Include initial limited version of libpam
This includes libpam.so and other parts of pam into the image.  However,
it does **not** integrate it with busybox or anything else.  That can
come later if desired.

This change includes enough of pam to allow python-pam to login (and
likely other integrations to work).

Additionally, the /etc/pam.d rules are very limited.  Again, this can be
coordinated/done in the future.

Inspired by:
https://github.com/LibreELEC/LibreELEC.tv/tree/221dceb8b7203109c3ea7c12ed4212f3822911d0/packages/security/Linux-PAM
(really old version of linux-pam in LibreELEC) and https://github.com/maximeh/buildroot/blob/master/package/linux-pam/linux-pam.mk (buildroot version of linux-pam)